refactor(ses): extract the configuration-set domain into SesConfigurationSetService
